refactor(grading): split trace_checks.py into constraints/ + evaluator package
- New tolokaforge/core/grading/trace_checks/ package: 6 support modules (truth, matcher, bindings, resolver, dispatch, evaluator) + constraints/ subpackage with 4 cluster modules (presence, ordering, windows, logical) + 2 __init__.py re-export shells
- 10 constraint operators grouped by cluster (each cluster's private helpers travel with consumers): presence.py (present, absent); ordering.py (before, immediately_before, count); windows.py (absent_before, absent_between); logical.py (all_of, any_of, negate)
- Delete trace_checks.py flat module (1659 LOC); package max file 445 LOC
- __init__.py re-exports all public + 7 private-underscore names for zero test-import edits
- _all_of/_any_of/_negate resolve _evaluate recursion via lazy function-body import (matches existing _operator_holds → plugin_registry pattern)
- test_trace_checks_constraints.py:617 AST audit retarget to evaluator.__file__ (function-local import; module-scope alias preserved)
- docs/GRADING.md:1766 path reference update
- Behaviour byte-identical: 486 trace-check tests + byte-parity 25/25 green

Closes #1352
