--- tolokaforge/core/grading/trace_checks.py
@@ -1,1659 +1,0 @@
-"""Scoring a pack's trace checks against the trial's event timeline.
-
-Substrate-neutral and pure — no services, no I/O — over the timeline both
-substrates build, so a constraint reaches the same verdict whichever substrate
-grades the trial.
-
-:func:`select_events` is the **only** function that resolves a matcher, and it
-resolves one under an environment of bound values. A constraint reads events
-through it and nothing else, so a constraint declaring a ``bind`` is the whole
-existing evaluator run once per candidate assignment rather than a second
-evaluator. :func:`evaluate_trace_checks` folds the constraint verdicts into the
-component score.
-
-Both layers are three-valued. An event either definitely matches, definitely does
-not, or cannot be decided because the evidence a predicate reads was never
-recorded — and the third case is a state the timeline reaches routinely, on every
-bundle re-graded without its tool-call record. Collapsing it into "did not match"
-would satisfy every negative constraint in the agent's favour, which
-``docs/GRADING.md`` G4 names as the hazard to avoid. A constraint is therefore
-decided only when every completion of the undecidable evidence agrees, and an
-undecided constraint is a failing sub-check that names what could not be read.
-
-The result carries the author keys this evaluation decomposed, per constraint
-kind, for the runner's accounted-keys ledger. Recording them here rather than in
-the runner phase is what makes the account honest: a kind the evaluation never
-reaches — one nested inside a composite the walk stopped descending into, say —
-is never recorded, where a runner-side walk of the *config* would report it as
-evaluated whatever the evaluator did.
-
-The authored vocabulary is documented in ``docs/GRADING.md`` § "Trace Checks".
-"""
-
-from __future__ import annotations
-
-import itertools
-import operator
-import re
-from collections.abc import Callable, Iterable, Mapping, Sequence
-from dataclasses import dataclass
-from enum import Enum
-from typing import Any
-
-from tolokaforge.core.grading.key_manifest import (
-    EVALUATED,
-    NO_TIMELINE_EVENTS_SKIP,
-    TRACE_ALTERNATIVES_KEY,
-    TRACE_CONSTRAINT_KEY_BY_KIND,
-    TRACE_CONSTRAINTS_KEY,
-    UNBOUND_BINDING_SKIP,
-)
-from tolokaforge.core.grading.predicates import ever_satisfiable, json_type_of
-from tolokaforge.core.grading.trace_timeline import (
-    TraceEvent,
-    TraceEventKind,
-    TrialTimeline,
-)
-from tolokaforge.core.models import (
-    AbsentBeforeConstraint,
-    AbsentBetweenConstraint,
-    AbsentConstraint,
-    AdjacencyView,
-    AnchorQuantifier,
-    BeforeConstraint,
-    BoundValue,
-    CountConstraint,
-    ImmediatelyBeforeConstraint,
-    KeyAccountingRecord,
-    MatcherSide,
-    OnMissing,
-    OnUnbound,
-    PresentConstraint,
-    Quantifier,
-    TraceBinding,
-    TraceChecksConfig,
-    TraceChecksResult,
-    TraceConstraint,
-    TraceConstraintExpr,
-    TraceConstraintKind,
-    TraceConstraintResult,
-    TraceConstraintSeverity,
-    TraceMatcher,
-    TracePathResult,
-    TurnWindow,
-    ValuePredicate,
-)
-
-# ``plugin_registry`` is imported lazily inside the two functions below that
-# use it (``_operator_holds`` + ``_binding_operator_names``) — see their bodies.
-
-__all__ = ["MatcherOutcome", "evaluate_trace_checks", "select_events"]
-
-
-class _Makeability(str, Enum):
-    """How one binding comparison read on one event.
-
-    Three-valued rather than two, because a comparison over a value no JSON type
-    names is evidence neither way, and reading it as ``MADE`` would let one call
-    that omitted the argument silence a report a sibling call earned.
-    """
-
-    MADE = "made"
-    UNMAKEABLE = "unmakeable"
-    NEITHER = "neither"
-
-
-@dataclass(frozen=True)
-class _ComparisonRecord:
-    """One binding reference of one matcher, as it read on one candidate event."""
-
-    event: TraceEvent
-    field: str
-    operator: str
-    binding: str
-    bound: Any
-    state: _Makeability
-
-    __hash__ = None
-    """Unhashable, like the :class:`TraceEvent` every record embeds — without this,
-    the generated hash could only raise, naming ``TraceEvent`` rather than the
-    record."""
-
-    @property
-    def reference(self) -> tuple[str, str, str]:
-        """What the record is a reading *of*, which repeats across the candidates."""
-        return (self.field, self.operator, self.binding)
-
-
-@dataclass(frozen=True)
-class MatcherOutcome:
-    """What one matcher resolved to on one timeline.
-
-    ``matched`` holds the events that match whatever the missing evidence turns out
-    to be, in ascending ``position``. ``undecidable`` holds the events whose every
-    other predicate passes but whose record-only evidence the timeline does not
-    carry, so some completion of the record would match them and another would not.
-
-    A constraint is decided when every completion of ``undecidable`` yields the same
-    verdict, and indeterminate otherwise — so a caller reads both tuples, never
-    ``matched`` alone.
-
-    ``comparisons`` holds how each binding reference the matcher declares read on
-    each event that was a candidate for it — the events every *other* reading of the
-    matcher admits, so a comparison speaks only for the events it was actually read
-    on.
-    """
-
-    matched: tuple[TraceEvent, ...]
-    undecidable: tuple[TraceEvent, ...]
-    unreadable_fields: tuple[str, ...]
-    comparisons: tuple[_ComparisonRecord, ...]
-
-    @property
-    def unmakeable_comparisons(self) -> tuple[str, ...]:
-        """The comparisons no candidate could make, as the sentences the grade prints.
-
-        A reference is reported when at least one candidate could not make its
-        comparison and none of them made it, which is the printed sentence's own
-        truth: "the comparison was not made" must not be printed on a trajectory
-        where some candidate made it. A candidate that read no comparison at all —
-        an argument it did not carry — neither reports nor silences, since a call
-        that omitted the argument is no evidence that the reference is reachable.
-        """
-        by_reference: dict[tuple[str, str, str], list[_ComparisonRecord]] = {}
-        for record in self.comparisons:
-            by_reference.setdefault(record.reference, []).append(record)
-        return tuple(
-            _unmakeable_message(records[0])
-            for records in by_reference.values()
-            if any(record.state is _Makeability.UNMAKEABLE for record in records)
-            and all(record.state is not _Makeability.MADE for record in records)
-        )
-
-    @property
-    def indeterminate_reason(self) -> str | None:
-        """Which evidence is missing and where, or ``None`` when the matcher is decided.
-
-        A property rather than a field: non-``None`` exactly when ``undecidable`` is
-        non-empty is the contract callers branch on, and deriving it makes the two
-        unable to disagree.
-        """
-        if not self.undecidable:
-            return None
-        return "the matcher cannot be decided: " + _missing_evidence(
-            self.unreadable_fields, (event.position for event in self.undecidable)
-        )
-
-
-def _missing_evidence(fields: Iterable[str], positions: Iterable[int]) -> str:
-    """Which record-only evidence the trial does not carry, and where.
-
-    One phrasing for both places evidence goes missing — a predicate that could not
-    be read and an extraction that could not be read — so the two cannot drift into
-    two vocabularies for the same gap.
-    """
-    where = sorted(positions)
-    label = "position" if len(where) == 1 else "positions"
-    return (
-        f"the trial records no {' or '.join(sorted(fields))} at "
-        f"{label} {', '.join(str(position) for position in where)}"
-    )
-
-
-class _Truth(str, Enum):
-    """Whether an event matches, given evidence that may be missing.
-
-    ``WITHHELD`` names the author-opted-out verdict :class:`OnMissing.WITHHOLD`
-    yields on an unmatched anchor. It is contagious under composites — an
-    ``all_of`` or ``any_of`` a withheld branch enters withholds too, unless a
-    definite verdict beats it (``FALSE`` under conjunction, ``TRUE`` under
-    disjunction), so an author writing ``all_of: [genuinely_failing,
-    withheld_thing]`` still learns which branch failed.
-    """
-
-    TRUE = "true"
-    FALSE = "false"
-    UNKNOWN = "unknown"
-    WITHHELD = "withheld"
-
-
-def select_events(
-    timeline: TrialTimeline, matcher: TraceMatcher, bindings: Mapping[str, Any]
-) -> MatcherOutcome:
-    """Resolve ``matcher`` against ``timeline`` under ``bindings``.
-
-    ``kind`` selects the event class and nothing is inferred from which predicates
-    are present. A ``tool_call`` matcher reads ``status`` and ``result`` from the
-    result paired to it by ``call_id`` — the call event itself carries neither — so
-    "a failed call to X with argument Y" is one matcher rather than two.
-
-    A predicate over a ``None`` field is unmatched rather than vacuously true, and
-    where that ``None`` means "only the tool-call record could have said, and it did
-    not" the event is undecidable instead.
-
-    ``bindings`` is the environment the constraint's binder produced for one
-    candidate assignment — the required third argument, so every call site says what
-    it resolves under, and empty for a matcher referencing nothing.
-    """
-    results = _results_by_call_id(timeline)
-    matched: list[TraceEvent] = []
-    undecidable: list[TraceEvent] = []
-    unreadable: set[str] = set()
-    comparisons: list[_ComparisonRecord] = []
-    for event in timeline.events:
-        if event.kind is not matcher.kind:
-            continue
-        truth, missing, records = _resolve(matcher, event, results, bindings)
-        comparisons.extend(records)
-        if truth is _Truth.TRUE:
-            matched.append(event)
-        elif truth is _Truth.UNKNOWN:
-            undecidable.append(event)
-            unreadable |= missing
-    return MatcherOutcome(
-        tuple(matched), tuple(undecidable), tuple(sorted(unreadable)), tuple(comparisons)
-    )
-
-
-# Fields only the tool-call record supplies. ``None`` in one of them is the
-# difference between "the agent did not do that" and "nobody wrote down what
-# happened", and a matcher that cannot tell them apart passes every negative
-# constraint on a re-graded bundle.
-_RECORD_ONLY_FIELDS = frozenset({"executor", "status"})
-
-
-def _unreadable_when_none(outcome: TraceEvent | None) -> frozenset[str]:
-    """The fields whose ``None`` on this event is missing evidence, not an answer.
-
-    With no outcome event at all — a call the trial never recorded a result for —
-    the text that call would have returned is unrecorded rather than empty, so a
-    ``result`` reading is as undecidable there as a ``status`` one. An outcome that
-    exists and carries no ``result`` is a definite absence instead.
-
-    The one rule for both readings of a field: a matcher's predicate over it and a
-    binding's extraction out of it.
-    """
-    if outcome is not None:
-        return _RECORD_ONLY_FIELDS
-    return _RECORD_ONLY_FIELDS | {"result"}
-
-
-def _resolve(
-    matcher: TraceMatcher,
-    event: TraceEvent,
-    results: Mapping[str, TraceEvent],
-    bindings: Mapping[str, Any],
-) -> tuple[_Truth, frozenset[str], list[_ComparisonRecord]]:
-    """What this event decides, what it left unread, and what it compared.
-
-    The verdict is the Kleene conjunction over the matcher's predicates. A
-    definitely-failing predicate decides the event whatever the missing evidence
-    would have said, so it wins over undecidability however the two are ordered —
-    which is what keeps an unexecuted call to a tool the matcher does not name out
-    of the undecidable set. A reading whose comparison could not be made is one of
-    those definite failures: ``40 == "report.md"`` is false.
-
-    A comparison is recorded on this event only where every *other* reading admits
-    it, so an event some other predicate definitely rejects speaks for no comparison
-    — which is what keeps a call to a tool the matcher does not name out of the
-    candidate set. A reading that made no comparison of its own rejects nothing,
-    whether it was refused or read a value no JSON type names: two bad references on
-    one matcher would otherwise empty each other's candidate set, and an author who
-    wrote two of them would be told about neither.
-    """
-    outcome = _outcome_of(event, results)
-    unreadable_when_none = _unreadable_when_none(outcome)
-    readings = _predicate_readings(matcher, event, outcome)
-    records = [
-        _comparison_records(field, None if value is _MISSING else value, predicate, bindings, event)
-        for field, value, predicate in readings
-    ]
-    unreadable = {
-        field for field, value, _ in readings if value is None and field in unreadable_when_none
-    }
-    failing = {
-        index
-        for index, (field, value, predicate) in enumerate(readings)
-        if not (value is None and field in unreadable_when_none)
-        and not _predicate_holds(predicate, value, bindings)
-    }
-    rejecting = {
-        index
-        for index in failing
-        if all(record.state is _Makeability.MADE for record in records[index])
-    }
-    candidate_records = [
-        record for index, found in enumerate(records) if not rejecting - {index} for record in found
-    ]
-    if failing:
-        return _Truth.FALSE, frozenset(), candidate_records
-    if unreadable:
-        return _Truth.UNKNOWN, frozenset(unreadable), candidate_records
-    return _Truth.TRUE, frozenset(), candidate_records
-
-
-def _comparison_records(
-    field: str,
-    value: Any,
-    predicate: ValuePredicate,
-    bindings: Mapping[str, Any],
-    event: TraceEvent,
-) -> list[_ComparisonRecord]:
-    """How each binding reference on this reading compared on this one event.
-
-    A predicate is a conjunction, so both references are read: an author who wrote
-    two of them made two mistakes and is owed both.
-    """
-    references = [(name, getattr(predicate, name)) for name in _binding_operator_names()]
-    return [
-        _ComparisonRecord(
-            event=event,
-            field=field,
-            operator=name,
-            binding=bound_name,
-            bound=bindings[bound_name],
-            state=_makeability(name, value, bindings[bound_name]),
-        )
-        for name, bound_name in references
-        if bound_name is not None
-    ]
-
-
-def _makeability(operator: str, value: Any, bound: Any) -> _Makeability:
-    """Whether this pair of runtime values could have satisfied the comparison.
-
-    The question is asked of both operands' JSON types, over
-    :func:`~tolokaforge.core.grading.predicates.ever_satisfiable`'s per-operator
-    table — which pairs are false on every trajectory is a property of the pair and
-    not of which side holds the text, so a text binding read against a natively-typed
-    field gets the same answer as the reverse. A pair the table refuses is
-    indistinguishable from an agent failure to everything that reads the score, and
-    so recorded as a comparison rather than folded in as one.
-
-    A side no JSON type names — an absent argument, a JSON ``null`` — makes the
-    reading ``NEITHER``: the comparison was not made there, and that says nothing
-    about whether the reference is reachable. That pre-gate is this caller's own, as
-    ``ever_satisfiable`` fails open on a type it cannot name.
-    """
-    held_type, bound_type = json_type_of(value), json_type_of(bound)
-    if held_type is None or bound_type is None:
-        return _Makeability.NEITHER
-    if ever_satisfiable(operator, held_type, bound_type):
-        return _Makeability.MADE
-    return _Makeability.UNMAKEABLE
-
-
-def _unmakeable_message(record: _ComparisonRecord) -> str:
-    """What the grade says about a reference no candidate could compare.
-
-    True of the candidate set as a whole rather than of any one event, so it names
-    the binding, the value it holds and that value's JSON type, the field and the
-    operator — and no held type, which differs across the candidates the sentence
-    speaks for. The remedy is one the authoring gate's own rules accept: a capture is
-    text only where the field beneath it holds text.
-    """
-    return (
-        f"the {record.field} comparison was not made: binding {record.binding!r} holds "
-        f"{record.bound!r}, a JSON {json_type_of(record.bound)}, and no candidate carried a "
-        f"value at that field which {record.operator} can ever satisfy against it — two JSON "
-        "types the operator cannot pair are false on every trajectory, whichever of the two "
-        "holds the text. Reference the binding from an args predicate whose arguments the "
-        "tools type the same way, or extract a regex capture off a field that holds text"
-    )
-
-
-def _predicate_readings(
-    matcher: TraceMatcher, event: TraceEvent, outcome: TraceEvent | None
-) -> list[tuple[str, Any, ValuePredicate]]:
-    """Every declared predicate paired with the value it reads on this event."""
-    declared = [
-        ("tool", event.tool_name, matcher.tool),
-        ("executor", event.executor, matcher.executor),
-        ("status", outcome.status if outcome is not None else None, matcher.status),
-        ("result", outcome.result if outcome is not None else None, matcher.result),
-        ("text", event.text, matcher.text),
-    ]
-    readings = [
-        (field, value, predicate) for field, value, predicate in declared if predicate is not None
-    ]
-    readings.extend(
-        (f"args.{path}", _argument_at(event.arguments, path), predicate)
-        for path, predicate in (matcher.args or {}).items()
-    )
-    return readings
-
-
-def _outcome_of(event: TraceEvent, results: Mapping[str, TraceEvent]) -> TraceEvent | None:
-    """The event a call's outcome is read from: the result paired to it, or itself."""
-    if event.kind is not TraceEventKind.TOOL_CALL:
-        return event
-    return results.get(event.call_id or "")
-
-
-def _results_by_call_id(timeline: TrialTimeline) -> dict[str, TraceEvent]:
-    return {
-        event.call_id: event
-        for event in timeline.events
-        if event.kind is TraceEventKind.TOOL_RESULT and event.call_id is not None
-    }
-
-
-# What an argument path resolves to where the key was never sent — distinct from
-# JSON ``null``, which resolves to ``None``. Only ``omitted`` reads the difference
-# between the two; every other operator receives ``None`` for both. The sentinel
-# never leaves this module: ``_operator_holds`` reads it directly for ``is_null`` /
-# ``omitted`` and collapses it to ``None`` for every other operator, ``_resolve``
-# collapses it at the tuple boundary before ``_comparison_records`` — the one
-# transit path that would otherwise reach ``json_type_of`` — and
-# ``_binder_reading`` collapses it at the extraction boundary.
-_MISSING: Any = object()
-
-
-def _argument_at(arguments: Mapping[str, Any] | None, path: str) -> Any:
-    """The value a dotted argument path addresses.
-
-    Returns the :data:`_MISSING` sentinel where the path does not resolve — an
-    ancestor that is not a mapping, or a segment absent from the mapping that
-    carries it. Every operator but ``omitted`` reads the sentinel as ``None`` at
-    :func:`_operator_holds`'s dispatch gate, so the key-absent path becomes the
-    same reading a JSON ``null`` already had. ``omitted`` is what makes the two
-    tellable apart, and reads the sentinel directly.
-    """
-    value: Any = arguments
-    for segment in path.split("."):
-        if not isinstance(value, Mapping) or segment not in value:
-            return _MISSING
-        value = value[segment]
-    return value
-
-
-def _predicate_holds(predicate: ValuePredicate, value: Any, bindings: Mapping[str, Any]) -> bool:
-    """Whether every operator the predicate declares holds — it is their conjunction.
-
-    A predicate declaring no operator is rejected at load, so the conjunction is
-    never over the empty set and never vacuously true.
-    """
-    return all(
-        _operator_holds(name, value, getattr(predicate, name), bindings)
-        for name in predicate.declared_operators()
-    )
-
-
-def _operator_holds(name: str, value: Any, expected: Any, bindings: Mapping[str, Any]) -> bool:
-    """One operator over one value.
-
-    ``is_null`` reads whether the field held an explicit JSON ``null``, and
-    ``omitted`` whether the key was never sent — the pair whose meaning turns on
-    the argument's state before the reading. Both are special-cased ahead of the
-    entry-point dispatch: the :data:`_MISSING` sentinel that separates the two
-    is private to this module, so the seam cannot answer for them; their
-    registered callables are stubs kept only to keep the frozenset and the
-    entry-point registry in lockstep. Every other operator reads a ``_MISSING``
-    as ``None`` — so a JSON ``null`` and an absent key collapse to one reading
-    at the seam and the sentinel never reaches a registered callable.
-
-    Only ``exists`` reads a ``None``. Every other operator is false there rather
-    than answering about a value the trial does not have — ``not_equals`` against
-    an absent argument would otherwise hold, which is the vacuous truth the
-    timeline contract forbids. The gate is a dispatch invariant declared once
-    here rather than repeated inside every registered operator.
-
-    ``name`` resolves through the ``tolokaforge.trace_check_operators``
-    entry-point group — the only dispatch table this evaluator reads.
-    """
-    if name == "is_null":
-        return (value is None) is expected
-    if name == "omitted":
-        return (value is _MISSING) is expected
-    if value is _MISSING:
-        value = None
-    if value is None and name != "exists":
-        return False
-    from tolokaforge.core.plugin_registry import load_trace_check_operator
-
-    op = load_trace_check_operator(name)
-    return op(value, expected, bindings)
-
-
-def _binding_operator_names() -> list[str]:
-    """Registered operator names whose semantics substitute a bound value.
-
-    Materialised from the entry-point registry, filtered by the ``_binding``
-    suffix — the sole marker for binding operators (ADR-0040). The discovery
-    scan is cached in ``plugin_registry``, so a per-call filter is O(N) over
-    the registry size (shipped defaults plus downstream) and does not fire
-    the loader.
-    """
-    from tolokaforge.core.plugin_registry import (
-        TRACE_CHECK_OPERATORS_GROUP,
-        discover_entry_points,
-    )
-
-    return sorted(
-        n for n in discover_entry_points(TRACE_CHECK_OPERATORS_GROUP) if n.endswith("_binding")
-    )
-
-
-def evaluate_trace_checks(timeline: TrialTimeline, config: TraceChecksConfig) -> TraceChecksResult:
-    """Score ``config``'s constraints against ``timeline``.
-
-    A block declaring ``alternatives`` is scored once per route, over the shared
-    constraints and that route's own, and the component is the best route's score
-    with the winner named. A route is scored as a whole rather than step by step,
-    so half of one route plus half of another beats neither.
-
-    A timeline carrying neither a conversational turn nor a tool call is the trial
-    that left no trace of itself: no constraint is evaluated and every declared
-    kind is accounted as skipped, because every constraint would otherwise score
-    against evidence the trial does not carry. A caller reads ``constraints`` to
-    tell the two apart — empty is the trial that left no trace.
-    """
-    if not timeline.events:
-        return TraceChecksResult(
-            accounted_keys=_accounting(config, _declared_kinds(config), NO_TIMELINE_EVENTS_SKIP)
-        )
-    ledger = _KindLedger()
-    shared = [
-        _evaluate_constraint(timeline, constraint, ledger) for constraint in config.constraints
-    ]
-    if config.alternatives is None:
-        return _component(config, _decision_set(shared), ledger, winner_id="", paths=[])
-    routes = {
-        path.id: _decision_set(
-            shared + [_evaluate_constraint(timeline, item, ledger) for item in path.constraints]
-        )
-        for path in config.alternatives
-    }
-    winner_id = max(routes, key=lambda path_id: _precedence(routes[path_id]))
-    paths = [
-        TracePathResult(id=path_id, score=route.score, gate_failed=bool(route.failed_gate_ids))
-        for path_id, route in routes.items()
-    ]
-    return _component(config, routes[winner_id], ledger, winner_id=winner_id, paths=paths)
-
-
-class _KindLedger:
-    """Which constraint kinds the evaluation walked, and which it never entered.
-
-    A kind reaches ``visited`` only by carrying a verdict — its ``require`` tree was
-    evaluated, or its constraint scored without entering one because the binder
-    yielded no assignment. A walk of the *config* would report a kind as evaluated
-    whatever the evaluator did, which is the accounting dishonesty this module exists
-    to avoid, and filing a kind the grade *failed the trial on* as skipped is the
-    same dishonesty pointing the other way.
-
-    ``unevaluated`` holds the kinds of the constraints whose ``require`` tree was
-    never entered, a state a binding that selected nothing reaches. The block files a
-    skip for those no constraint scored — so under a skipped tree that is its nested
-    kinds, the top-level one having taken the constraint's own verdict.
-    """
-
-    def __init__(self) -> None:
-        self.visited: set[TraceConstraintKind] = set()
-        self.unevaluated: set[TraceConstraintKind] = set()
-
-    def skipped_kinds(self) -> set[TraceConstraintKind]:
-        """The kinds no constraint evaluated, which the block accounts as skipped."""
-        return self.unevaluated - self.visited
-
-
-@dataclass(frozen=True)
-class _DecisionSet:
-    """One route's verdicts — the shared constraints and that route's own — folded.
-
-    ``score`` is the route's own normalised score, before a gate zeroes anything:
-    a gate shuts the component, never the number a route earned, and the argmax
-    runs over these rather than over the zeroed values so that a route cannot
-    escape its own gate by scoring below a clean sibling.
-    """
-
-    results: list[TraceConstraintResult]
-    score: float
-    failed_gate_ids: list[str]
-
-
-def _precedence(route: _DecisionSet) -> tuple[float, bool]:
-    """What the argmax orders routes by: the score, then whether a gate shut.
-
-    ``max`` keeps the first of equal keys and the routes are in declaration order,
-    so a tie between clean routes is broken by which one the author wrote first.
-    A route whose gate failed wins a tie outright: it can only ever shut the
-    component and never rescue it, so the trial's verdict does not turn on where
-    in the file the two routes were written.
-    """
-    return (route.score, bool(route.failed_gate_ids))
-
-
-def _decision_set(results: list[TraceConstraintResult]) -> _DecisionSet:
-    """``results`` scored over its non-gate, non-withheld members, or collapsed to the gate verdict.
-
-    A gate enters neither the numerator nor the denominator, so a decision set whose
-    every member is a gate has no weighted average to take — and an author who wrote
-    only gates asked for a defined verdict, not an empty sum. It scores the gates'
-    own verdict, which is what :func:`~tolokaforge.core.grading.rubric.aggregate_rubric`
-    already returns for an all-required rubric. The collapse lives here rather than inside
-    :func:`_weighted_fraction`, whose denominator its callers keep positive, and it
-    is not conditional on ``alternatives``: a flat block of nothing but gates
-    reaches it too.
-
-    A withheld constraint is out of the decision on both sides: excluded from the
-    numerator and denominator of the weighted fraction, from the all-gates collapse,
-    and from ``failed_gate_ids`` — a withheld gate is neither passing nor failing.
-    """
-    scored = [
-        item
-        for item in results
-        if item.severity is not TraceConstraintSeverity.GATE and not item.withheld
-    ]
-    decided = [item for item in results if not item.withheld]
-    return _DecisionSet(
-        results=results,
-        score=(
-            _weighted_fraction(scored)
-            if scored
-            else (1.0 if all(item.passed for item in decided) else 0.0)
-        ),
-        failed_gate_ids=[
-            item.id
-            for item in results
-            if item.severity is TraceConstraintSeverity.GATE
-            and not item.passed
-            and not item.withheld
-        ],
-    )
-
-
-def _component(
-    config: TraceChecksConfig,
-    winner: _DecisionSet,
-    ledger: _KindLedger,
-    *,
-    winner_id: str,
-    paths: list[TracePathResult],
-) -> TraceChecksResult:
-    """The winning decision set as the component both substrates read.
-
-    A tripped gate zeroes the score here rather than at each substrate's
-    integration, because ``TraceChecksResult.score`` *is* the component both
-    :meth:`~tolokaforge.core.grading.combine.GradingEngine.grade_trajectory` and the
-    runner's ``GradeTrial`` assign — the same act ``GradeTrial`` performs on the
-    judge's aggregate when a required criterion fails, done once here for both.
-    """
-    return TraceChecksResult(
-        passed=all(item.passed for item in winner.results if not item.withheld),
-        score=0.0 if winner.failed_gate_ids else winner.score,
-        constraints=winner.results,
-        winning_path=winner_id,
-        gate_failed=bool(winner.failed_gate_ids),
-        failed_gate_ids=winner.failed_gate_ids,
-        paths=paths,
-        # The skips are filed per kind and never against the block's own keys: a
-        # constraint that bound nothing leaves the kinds nested under its ``require``
-        # tree unaccounted, while the block itself was evaluated, and a kind any
-        # constraint scored keeps that record because ``skipped_kinds`` subtracts it.
-        accounted_keys=_accounting(config, ledger.visited, EVALUATED)
-        | {
-            TRACE_CONSTRAINT_KEY_BY_KIND[kind]: UNBOUND_BINDING_SKIP
-            for kind in ledger.skipped_kinds()
-        },
-    )
-
-
-def _accounting(
-    config: TraceChecksConfig,
-    kinds: Iterable[TraceConstraintKind],
-    record: KeyAccountingRecord,
-) -> dict[str, KeyAccountingRecord]:
-    """``record`` filed against the keys the block declares and each kind's own."""
-    alternatives = {TRACE_ALTERNATIVES_KEY: record} if config.alternatives is not None else {}
-    return {
-        TRACE_CONSTRAINTS_KEY: record,
-        **alternatives,
-        **{TRACE_CONSTRAINT_KEY_BY_KIND[kind]: record for kind in kinds},
-    }
-
-
-def _declared_kinds(config: TraceChecksConfig) -> set[TraceConstraintKind]:
-    """Every kind the block declares, shared or inside a path, nesting included."""
-    declared = [
-        *config.constraints,
-        *(item for path in config.alternatives or () for item in path.constraints),
-    ]
-    return {kind for item in declared for kind in item.require.kinds_in_tree()}
-
-
-def _weighted_fraction(results: Sequence[TraceConstraintResult]) -> float:
-    """``Σ(weight · passed) / Σ(weight)`` over the scored constraints of a decision set.
-
-    Every weight is positive and the only caller hands this a non-empty set — a
-    decision set with no scored member at all is answered by the gate collapse in
-    :func:`_decision_set` — so the denominator is positive by construction. There is
-    deliberately no zero-denominator branch: any score it returned would be a
-    convention no author chose, and keeping the empty case out is what removes the
-    need for one.
-    """
-    total = sum(result.weight for result in results)
-    earned = sum(result.weight for result in results if result.passed)
-    return earned / total
-
-
-def _evaluate_constraint(
-    timeline: TrialTimeline, constraint: TraceConstraint, ledger: _KindLedger
-) -> TraceConstraintResult:
-    """One constraint's verdict, its ``require`` tree read once per bound candidate.
-
-    Quantification over the candidates is outermost and universal: the constraint
-    holds when it holds under every assignment its binder yields. So ``negate``
-    inside a bound constraint reads "no candidate satisfies", not "not every
-    candidate does", and a constraint with one candidate scores exactly as the same
-    constraint with that value written as a literal.
-
-    The candidate set is itself three-valued, so the fold runs over the completions
-    of the undecidable candidates rather than over one set — and a candidate whose
-    value the trial does not record is read as ``UNKNOWN`` without entering the
-    ``require`` tree, since there is no environment to enter it under.
-    """
-    kind = constraint.require.declared_kind()
-    candidates = _candidates(timeline, constraint)
-    on_missing = constraint.on_missing or OnMissing.FAIL
-    definite = [
-        _read_under(timeline, constraint, environment, ledger.visited, on_missing)
-        for environment in candidates.definite
-    ]
-    possible = [
-        _read_under(timeline, constraint, environment, ledger.visited, on_missing)
-        for environment in candidates.undecidable
-    ]
-    if not definite and not possible:
-        # The top-level kind is scored whatever the binder yielded — this branch
-        # returns a verdict under it — so only the kinds nested beneath it went
-        # unreached. Filing the scored one as a skip would report a kind as
-        # unevaluated in the same grade that fails the trial on it.
-        ledger.visited.add(kind)
-        ledger.unevaluated |= constraint.require.kinds_in_tree()
-        if not candidates.unnamed:
-            return _unbound_result(constraint, kind, candidates)
-    readings = definite + possible
-    unmakeable = list(dict.fromkeys(item for reading in readings for item in reading.unmakeable))
-    truth = (
-        _Truth.FALSE
-        if unmakeable
-        else _folded_truth(
-            [reading.truth for reading in definite],
-            [reading.truth for reading in possible]
-            + ([_Truth.UNKNOWN] if candidates.unnamed else []),
-            _unbound_truth(constraint.bind),
-        )
-    )
-    withheld = truth is _Truth.WITHHELD
-    return TraceConstraintResult(
-        id=constraint.id,
-        kind=kind,
-        passed=truth is _Truth.TRUE,
-        undecided=truth is _Truth.UNKNOWN,
-        withheld=withheld,
-        weight=constraint.weight,
-        severity=constraint.severity,
-        message=(
-            "; ".join(unmakeable)
-            if unmakeable
-            else _folded_message(readings, truth, kind, candidates.undetermined)
-        ),
-        matched_positions=(
-            [] if withheld else sorted({item for reading in readings for item in reading.positions})
-        ),
-    )
-
-
-def _folded_truth(
-    definite: Sequence[_Truth], undecidable: Sequence[_Truth], unbound: _Truth
-) -> _Truth:
-    """The verdict every completion of the candidate set agrees on, or ``UNKNOWN``.
-
-    Kleene AND over a set is its minimum under ``FALSE < UNKNOWN < TRUE``, so over
-    the completions ``D ∪ S`` for ``S ⊆ U`` the reachable verdicts are exactly those
-    of the empty ``S``, of each singleton, and of ``U`` itself — a minimum over any
-    larger ``S`` equals one of its members'. Enumerating only the two ends drops the
-    singletons, which is sound while the empty reading is vacuously true and wrong
-    the moment ``D`` is empty: there the empty reading is ``on_unbound``, and a
-    completion binding one satisfied candidate can hold where both ends fail. That
-    over-fail is the hazard :func:`_reachable_counts` already names one level down.
-    """
-    inside = _conjunction(definite)
-    empty_reading = inside if definite else unbound
-    if not undecidable:
-        return empty_reading
-    readings = {
-        empty_reading,
-        *(_conjunction((inside, truth)) for truth in undecidable),
-        _conjunction((inside, *undecidable)),
-    }
-    return readings.pop() if len(readings) == 1 else _Truth.UNKNOWN
-
-
-@dataclass(frozen=True)
-class _CandidateReading:
-    """What one constraint's ``require`` tree decided under one bound assignment."""
-
-    environment: Mapping[str, Any]
-    truth: _Truth
-    message: str
-    positions: list[int]
-    unmakeable: list[str]
-
-
-def _read_under(
-    timeline: TrialTimeline,
-    constraint: TraceConstraint,
-    environment: Mapping[str, Any],
-    visited: set[TraceConstraintKind],
-    on_missing: OnMissing,
-) -> _CandidateReading:
-    resolver = _Resolver(timeline, constraint.within, visited, environment)
-    truth = _evaluate(constraint.require, resolver, on_missing)
-    return _CandidateReading(
-        environment=environment,
-        truth=truth,
-        message=_message(truth, constraint.require.declared_kind(), resolver, on_missing),
-        positions=resolver.matched_positions(),
-        unmakeable=resolver.unmakeable_comparisons(),
-    )
-
-
-def _folded_message(
-    readings: list[_CandidateReading],
-    truth: _Truth,
-    kind: TraceConstraintKind,
-    undetermined: str,
-) -> str:
-    """What the grade says about a constraint folded over its candidates.
-
-    The sentence is the one the first candidate that reached the folded verdict
-    wrote — the reading whose truth *is* the folded truth, so a definite failure
-    beside an undecided candidate reports the failure rather than the doubt. The
-    assignments named beside it are **every** candidate that reached it. Naming them
-    is the only way an author learns which record failed to correlate: the per-kind
-    detail sentence is the same whichever value it failed under, so a bound
-    constraint without them reports a failure the author cannot act on. A constraint
-    that binds nothing names none, since its one reading holds the empty
-    environment.
-
-    ``undetermined`` is reported only where the completions of the candidate set
-    disagreed, as a matcher's own missing evidence is: where they agree the evidence
-    the trial lacks changed nothing an author can act on. It is the whole
-    explanation where no single candidate reached the verdict, which is the shape a
-    set whose completions disagree among themselves takes.
-    """
-    if truth is _Truth.TRUE:
-        return ""
-    clause = undetermined if truth is _Truth.UNKNOWN else ""
-    responsible = [item for item in readings if item.truth is truth]
-    if not responsible:
-        return f"{kind.value} cannot be decided — {clause}"
-    sentence = responsible[0].message + _naming(responsible, truth)
-    return f"{sentence}; {clause}" if clause else sentence
-
-
-def _naming(readings: list[_CandidateReading], truth: _Truth) -> str:
-    named = ", ".join(_assignment_text(item.environment) for item in readings if item.environment)
-    return f"; {_FOLD_PHRASE[truth]} {named}" if named else ""
-
-
-def _assignment_text(environment: Mapping[str, Any]) -> str:
-    """One assignment as an author reads it — every name it binds, parenthesised.
-
-    Parenthesised even at one name, so that a constraint binding two names reads
-    unambiguously when several assignments are listed side by side.
-    """
-    return "(" + ", ".join(f"{name}={value!r}" for name, value in sorted(environment.items())) + ")"
-
-
-# How a fold over candidates reads its own verdict, beside the values it reached it
-# under. Neutral about what the constraint asserted, because a ``negate`` fails
-# where its correlation *held*. ``TRUE`` is on no row: a passing constraint says
-# nothing at all.
-_FOLD_PHRASE: Mapping[_Truth, str] = {
-    _Truth.FALSE: "failed under",
-    _Truth.UNKNOWN: "undecided under",
-    _Truth.WITHHELD: "withheld under",
-}
-
-
-def _unbound_truth(bind: TraceBinding | None) -> _Truth:
-    """What the empty reading of the candidate set decides.
-
-    The universal reading is vacuously true over an empty candidate set, so the
-    author's ``on_unbound`` supplies the verdict instead — defaulting to a failure,
-    because a binder that never fired usually means the agent never did the thing
-    the constraint is about. A constraint declaring no binder never reads this: its
-    one candidate is the empty environment, so its set is never empty.
-    """
-    if bind is not None and bind.on_unbound is OnUnbound.PASS:
-        return _Truth.TRUE
-    return _Truth.FALSE
-
-
-def _unbound_result(
-    constraint: TraceConstraint, kind: TraceConstraintKind, candidates: _Candidates
-) -> TraceConstraintResult:
-    """The verdict of a bound constraint whose binder yielded no assignment at all."""
-    truth = _unbound_truth(constraint.bind)
-    passed = truth is _Truth.TRUE
-    return TraceConstraintResult(
-        id=constraint.id,
-        kind=kind,
-        passed=passed,
-        undecided=truth is _Truth.UNKNOWN,
-        weight=constraint.weight,
-        severity=constraint.severity,
-        message="" if passed else f"{kind.value} is unbound: {candidates.emptiness}",
-        matched_positions=[],
-    )
-
-
-@dataclass(frozen=True)
-class _Candidates:
-    """The assignments a constraint's binder yields, and what it left undetermined.
-
-    ``definite`` are the assignments the trial binds whatever its missing evidence
-    would have said; ``undecidable`` are the ones some completion of that evidence
-    binds and another does not. ``unnamed`` marks a binder event whose extracted
-    value the trial does not record at all — a candidate under some completion, with
-    no value this trial can name, so it decides nothing and leaves the fold undecided.
-
-    ``missing`` maps the position of each event that left the set undetermined to the
-    evidence the trial does not carry there, and holds only events that contributed:
-    one whose membership is unsettled and that carried no value to bind either way
-    changes no reading of the set.
-
-    ``emptiness`` separates the two ways a candidate set is empty — no event
-    selected, or events selected that carried no value to bind — because they are
-    different author mistakes and a single "no match" tells neither.
-    """
-
-    definite: list[Mapping[str, Any]]
-    undecidable: list[Mapping[str, Any]]
-    unnamed: bool
-    missing: Mapping[int, frozenset[str]]
-    emptiness: str
-
-    @property
-    def undetermined(self) -> str:
-        """Why the candidate set itself could not be determined, empty when it was.
-
-        Non-empty exactly where the binder left something open — an event whose
-        membership the trial cannot settle, or one whose value it does not record.
-        """
-        if not self.missing:
-            return ""
-        fields = frozenset[str]().union(*self.missing.values())
-        return "the candidate set cannot be determined: " + _missing_evidence(fields, self.missing)
-
-
-_UNBOUND_ENVIRONMENT: Mapping[str, Any] = {}
-"""What a constraint declaring no binder resolves its matchers under."""
-
-
-def _candidates(timeline: TrialTimeline, constraint: TraceConstraint) -> _Candidates:
-    """Every distinct assignment the binder yields, in the order its events occur.
-
-    The binder resolves through :func:`select_events` like every other matcher, so
-    it has an undecidable set of its own: an event whose membership the trial cannot
-    settle yields a candidate the constraint must be decided *without* assuming, and
-    one whose extraction reads unrecorded evidence yields a candidate with no value.
-    """
-    if constraint.bind is None:
-        return _Candidates([_UNBOUND_ENVIRONMENT], [], False, {}, "")
-    outcome = _restricted(
-        select_events(timeline, constraint.bind.match, _UNBOUND_ENVIRONMENT), constraint.within
-    )
-    results = _results_by_call_id(timeline)
-    bound: list[Mapping[str, Any]] = []
-    possible: list[Mapping[str, Any]] = []
-    missing: dict[int, frozenset[str]] = {}
-    unnamed = False
-    for event, settled in _selected(outcome):
-        reading = _bound_event(constraint.bind, event, results)
-        (bound if settled else possible).extend(reading.assignments)
-        unnamed = unnamed or bool(reading.unreadable)
-        membership = frozenset() if settled else frozenset(outcome.unreadable_fields)
-        # Only an event that contributed puts the set in doubt: one whose membership
-        # the binder could not settle and that carried no value to bind either way
-        # changes no reading of the set.
-        if reading.unreadable or (reading.assignments and not settled):
-            missing[event.position] = reading.unreadable | membership
-    definite = _distinct(bound)
-    undecidable = _distinct(possible, definite)
-    return _Candidates(
-        definite,
-        undecidable,
-        unnamed,
-        missing,
-        "" if definite or undecidable or unnamed else _emptiness(outcome),
-    )
-
-
-def _selected(outcome: MatcherOutcome) -> list[tuple[TraceEvent, bool]]:
-    """Every event the binder drew a candidate from, and whether it definitely did."""
-    return [
-        *((event, True) for event in outcome.matched),
-        *((event, False) for event in outcome.undecidable),
-    ]
-
-
-def _distinct(
-    assignments: Sequence[Mapping[str, Any]], already: Sequence[Mapping[str, Any]] = ()
-) -> list[Mapping[str, Any]]:
-    """``assignments`` with the repeats dropped, keeping the first of each.
-
-    Candidates are distinct *values*, not events: ten calls naming one record are
-    one reading of the ``require`` tree rather than ten identical ones, and the
-    values a failure names are the ones the author has to look at rather than that
-    list with duplicates in it. A linear scan rather than a set because an
-    assignment is a dict and a bound value may itself be a list or a dict — nothing
-    here is hashable.
-
-    ``already`` are assignments the caller has resolved ahead of these, and drops an
-    undecidable candidate a definite one already binds: a value the trial definitely
-    binds is in the set whatever the missing evidence says, so a second undecidable
-    copy of it enumerates a reading already read and names the value twice.
-    """
-    distinct = list(already)
-    kept: list[Mapping[str, Any]] = []
-    for assignment in assignments:
-        if not any(_same_assignment(assignment, seen) for seen in distinct):
-            distinct.append(assignment)
-            kept.append(assignment)
-    return kept
-
-
-def _same_assignment(left: Mapping[str, Any], right: Mapping[str, Any]) -> bool:
-    """Whether two assignments bind every name to the same value of the same type.
-
-    Type as well as equality, because ``True == 1`` in Python: equality alone would
-    collapse two values the arguments they came from distinguish, and the surviving
-    one would then decide the constraint for both. Every assignment one binder
-    yields carries the same names, so one side's are enough to compare over.
-    """
-    return all(type(left[name]) is type(right[name]) and left[name] == right[name] for name in left)
-
-
-def _emptiness(outcome: MatcherOutcome) -> str:
-    selected = len(outcome.matched) + len(outcome.undecidable)
-    if not selected:
-        return "the binding selected no event"
-    label = "event" if selected == 1 else "events"
-    return f"the binding selected {selected} {label}, none of which carried every value it extracts"
-
-
-@dataclass(frozen=True)
-class _BoundEvent:
-    """What one event the binder selected contributes to the candidate set.
-
-    ``unreadable`` names the evidence some completion of the record would have bound
-    a value out of and this trial does not carry. It is non-empty only where
-    ``assignments`` is empty, since a name reading nothing empties the product: the
-    event is a candidate whose value the trial cannot name, which is missing evidence
-    rather than an event carrying no value.
-    """
-
-    assignments: list[Mapping[str, Any]]
-    unreadable: frozenset[str]
-
-
-def _bound_event(
-    binding: TraceBinding, event: TraceEvent, results: Mapping[str, TraceEvent]
-) -> _BoundEvent:
-    """One assignment per element of the cross product of the names' extracted values.
-
-    A plain ``field`` reads one value and a ``pattern`` one per capture, so a name
-    reading nothing empties the product — the event binds no assignment at all
-    rather than an assignment with a hole in it. Which emptiness that is follows
-    :func:`_unreadable_when_none`, the same rule a predicate over the field reads:
-    an absent argument is an absent value, and a ``result`` on a call the trial
-    recorded no outcome for is evidence the trial does not carry.
-    """
-    outcome = _outcome_of(event, results)
-    names = sorted(binding.values)
-    extracted = [_extracted(binding.values[name], event, outcome) for name in names]
-    unreadable = _unreadable_when_none(outcome).intersection(
-        binding.values[name].head_segment()
-        for name, values in zip(names, extracted, strict=True)
-        if not values
-    )
-    return _BoundEvent(
-        [dict(zip(names, values, strict=True)) for values in itertools.product(*extracted)],
-        unreadable,
-    )
-
-
-def _extracted(bound: BoundValue, event: TraceEvent, outcome: TraceEvent | None) -> list[Any]:
-    """The values one name reads off one event, in the order the event carries them."""
-    value = _binder_reading(bound, event, outcome)
-    if value is None:
-        return []
-    if bound.pattern is None:
-        return [value]
-    if not isinstance(value, str):
-        return []
-    return [match.group(1) for match in re.finditer(bound.pattern, value)]
-
-
-def _binder_reading(bound: BoundValue, event: TraceEvent, outcome: TraceEvent | None) -> Any:
-    """The raw value the extraction addresses, before any capture narrows it.
-
-    An extraction reads a JSON ``null`` and an absent key as one condition — the
-    call that omitted the argument silences a report a sibling call earned, so
-    the :data:`_MISSING` sentinel :func:`_argument_at` returns is folded into
-    ``None`` at this boundary and never reaches a bound value.
-    """
-    head = bound.head_segment()
-    if head != "args":
-        return _BINDER_FIELDS[head](event, outcome)
-    _, _, path = bound.field.partition(".")
-    value = _argument_at(event.arguments, path) if path else event.arguments
-    return None if value is _MISSING else value
-
-
-# ``status`` and ``executor`` are on no row: a binding over a closed vocabulary of a
-# handful of members is rejected at load, so the extraction never reaches here.
-_BINDER_FIELDS: Mapping[str, Callable[[TraceEvent, TraceEvent | None], Any]] = {
-    "tool": lambda event, outcome: event.tool_name,
-    "text": lambda event, outcome: event.text,
-    "result": lambda event, outcome: outcome.result if outcome is not None else None,
-}
-
-
-@dataclass(frozen=True)
-class _Resolved:
-    """One matcher of one constraint, and what it selected.
-
-    ``anchor`` marks the positions whose emptiness is an *unmatched* constraint
-    rather than an answer: a `before` side that selected nothing leaves the
-    ordering unasked, while an `absent` matcher selecting nothing is the very thing
-    it asserts.
-    """
-
-    label: str
-    outcome: MatcherOutcome
-    anchor: bool
-
-
-class _Resolver:
-    """Resolves one constraint's matchers, remembering what each one found.
-
-    ``visited_kinds`` is the block's accumulator rather than this constraint's:
-    every expression the evaluation reaches adds its kind to it, and the block's
-    accounting is what the set holds once every constraint has been evaluated.
-    """
-
-    def __init__(
-        self,
-        timeline: TrialTimeline,
-        within: TurnWindow | None,
-        visited_kinds: set[TraceConstraintKind],
-        bindings: Mapping[str, Any],
-    ) -> None:
-        self.timeline = timeline
-        self.visited_kinds = visited_kinds
-        self._within = within
-        self._bindings = bindings
-        self._resolved: list[_Resolved] = []
-
-    def resolve(self, label: str, matcher: TraceMatcher, *, anchor: bool) -> MatcherOutcome:
-        outcome = _restricted(select_events(self.timeline, matcher, self._bindings), self._within)
-        self._resolved.append(_Resolved(label, outcome, anchor))
-        return outcome
-
-    def unmakeable_comparisons(self) -> list[str]:
-        """Every comparison a bound value's type put out of reach, in matcher order."""
-        return list(
-            dict.fromkeys(
-                message
-                for item in self._resolved
-                for message in item.outcome.unmakeable_comparisons
-            )
-        )
-
-    def matched_positions(self) -> list[int]:
-        return sorted({event.position for item in self._resolved for event in item.outcome.matched})
-
-    def undecided(self) -> list[str]:
-        """Why each matcher that could not be decided could not be decided."""
-        return [
-            f"{item.label}: {item.outcome.indeterminate_reason}"
-            for item in self._resolved
-            if item.outcome.indeterminate_reason is not None
-        ]
-
-    def unmatched_anchors(self) -> list[str]:
-        return [
-            item.label
-            for item in self._resolved
-            if item.anchor and not item.outcome.matched and not item.outcome.undecidable
-        ]
-
-    def empty_sides(self) -> list[str]:
-        """Every resolved matcher whose outcome carried no event, anchor or not.
-
-        Reads the withheld verdict on ``count`` — whose matcher is not an
-        anchor (its empty count is a defined verdict) but whose ``WITHHELD``
-        under ``on_missing: withhold`` still names the same side.
-        """
-        return [
-            item.label
-            for item in self._resolved
-            if not item.outcome.matched and not item.outcome.undecidable
-        ]
-
-
-def _restricted(outcome: MatcherOutcome, within: TurnWindow | None) -> MatcherOutcome:
-    """``outcome`` with every event outside the constraint's turn window dropped.
-
-    The window narrows what a matcher selects, not what the timeline contains, so
-    positions stay the trial's own and an adjacency view is still read over the
-    whole trial. It narrows the comparisons the same way, so a reference reduces
-    over the candidates the author's window admits and a call outside it neither
-    reports nor silences.
-    """
-    if within is None:
-        return outcome
-    matched = tuple(event for event in outcome.matched if _inside(event, within))
-    undecidable = tuple(event for event in outcome.undecidable if _inside(event, within))
-    return MatcherOutcome(
-        matched,
-        undecidable,
-        outcome.unreadable_fields if undecidable else (),
-        tuple(record for record in outcome.comparisons if _inside(record.event, within)),
-    )
-
-
-def _inside(event: TraceEvent, window: TurnWindow) -> bool:
-    below = window.first_turn is not None and event.turn_index < window.first_turn
-    above = window.last_turn is not None and event.turn_index > window.last_turn
-    return not below and not above
-
-
-def _evaluate(expr: TraceConstraintExpr, resolver: _Resolver, on_missing: OnMissing) -> _Truth:
-    kind = expr.declared_kind()
-    resolver.visited_kinds.add(kind)
-    return _HANDLERS[kind](getattr(expr, kind.value), resolver, on_missing)
-
-
-def _present(payload: PresentConstraint, resolver: _Resolver, on_missing: OnMissing) -> _Truth:
-    """No match is this constraint's own ``False``, unless the author opted to withhold.
-
-    The matcher is resolved as an anchor so the grade names which side selected
-    nothing. A matcher that yielded no candidate at all — no definite match and
-    no undecidable one — is what ``on_missing: withhold`` opts out of: the
-    verdict is ``WITHHELD`` and the check leaves numerator and denominator
-    alone. With a match or an undecidable candidate the reading falls through
-    to the count-based decision, where ``PASS`` and ``FAIL`` are refused at
-    load and ``WITHHOLD`` behaves as ``FAIL`` on any completion that reads
-    zero.
-    """
-    counts = _reachable_counts(resolver.resolve("match", payload.match, anchor=True))
-    if on_missing is OnMissing.WITHHOLD and counts == range(0, 1):
-        return _Truth.WITHHELD
-    return _decide((count > 0 for count in counts), on_missing)
-
-
-def _absent(payload: AbsentConstraint, resolver: _Resolver, on_missing: OnMissing) -> _Truth:
-    counts = _reachable_counts(resolver.resolve("match", payload.match, anchor=False))
-    return _decide((count == 0 for count in counts), on_missing)
-
-
-def _count(payload: CountConstraint, resolver: _Resolver, on_missing: OnMissing) -> _Truth:
-    """The count-based verdict, unless the matcher yielded nothing and the author withheld.
-
-    An empty outcome — no definite match, no undecidable one — is what
-    ``on_missing: withhold`` opts out of, and the ``min`` / ``max`` bounds are
-    irrelevant there: the withhold is on the absence of the anchor, not on
-    whether the count falls in bounds. With any candidate the reading falls
-    through to the bounds check.
-    """
-    counts = _reachable_counts(resolver.resolve("match", payload.match, anchor=False))
-    if on_missing is OnMissing.WITHHOLD and counts == range(0, 1):
-        return _Truth.WITHHELD
-    return _decide(
-        (_within_bounds(count, payload.min, payload.max) for count in counts), on_missing
-    )
-
-
-def _reachable_counts(outcome: MatcherOutcome) -> range:
-    """Every match count some completion of the undecidable evidence produces.
-
-    The whole interval, not its two ends: ``count: {min: 1, max: 1}`` over no
-    definite matches and two undecidables is satisfied by exactly one of them
-    matching, while neither end of ``[0, 2]`` satisfies it — so a rule reading only
-    the ends would answer "definitely false" where a completion says otherwise.
-    """
-    definite = len(outcome.matched)
-    return range(definite, definite + len(outcome.undecidable) + 1)
-
-
-def _within_bounds(count: int, minimum: int | None, maximum: int | None) -> bool:
-    return (minimum is None or count >= minimum) and (maximum is None or count <= maximum)
-
-
-def _before(payload: BeforeConstraint, resolver: _Resolver, on_missing: OnMissing) -> _Truth:
-    return _ordering(payload.left, payload.right, resolver, on_missing, operator.lt)
-
-
-def _immediately_before(
-    payload: ImmediatelyBeforeConstraint, resolver: _Resolver, on_missing: OnMissing
-) -> _Truth:
-    adjacent = _adjacency(resolver.timeline, payload.among)
-    return _ordering(payload.left, payload.right, resolver, on_missing, adjacent)
-
-
-def _ordering(
-    left: MatcherSide,
-    right: MatcherSide,
-    resolver: _Resolver,
-    on_missing: OnMissing,
-    relation: Callable[[int, int], bool],
-) -> _Truth:
-    lefts = _side_readings(resolver.resolve("left", left.match, anchor=True), left.quantifier)
-    rights = _side_readings(resolver.resolve("right", right.match, anchor=True), right.quantifier)
-    return _decide(
-        (
-            _relation_holds(chosen_left, left.quantifier, chosen_right, right.quantifier, relation)
-            for chosen_left in lefts
-            for chosen_right in rights
-        ),
-        on_missing,
-    )
-
-
-def _relation_holds(
-    left: _Reading,
-    left_quantifier: Quantifier,
-    right: _Reading,
-    right_quantifier: Quantifier,
-    relation: Callable[[int, int], bool],
-) -> bool | None:
-    """Whether the relation holds over this pair of readings, quantifier by quantifier.
-
-    ``first`` / ``last`` have already reduced their side to one position, so an
-    existential and a universal reading of it answer the same.
-    """
-    if not left or not right:
-        return None
-    return _quantified(
-        left_quantifier,
-        (
-            _quantified(right_quantifier, (relation(position, other) for other in right))
-            for position in left
-        ),
-    )
-
-
-def _quantified(quantifier: Quantifier, values: Iterable[bool]) -> bool:
-    return all(values) if quantifier is Quantifier.ALL else any(values)
-
-
-def _absent_before(
-    payload: AbsentBeforeConstraint, resolver: _Resolver, on_missing: OnMissing
-) -> _Truth:
-    anchors = _side_readings(
-        resolver.resolve("anchor", payload.anchor.match, anchor=True), payload.anchor.quantifier
-    )
-    forbidden = _side_readings(resolver.resolve("forbidden", payload.forbidden, anchor=False), None)
-    return _decide(
-        (
-            _nothing_inside(_prefix_window(anchor), blocked)
-            for anchor in anchors
-            for blocked in forbidden
-        ),
-        on_missing,
-    )
-
-
-def _absent_between(
-    payload: AbsentBetweenConstraint, resolver: _Resolver, on_missing: OnMissing
-) -> _Truth:
-    starts = _side_readings(
-        resolver.resolve("start", payload.start.match, anchor=True), payload.start.quantifier
-    )
-    ends = _side_readings(
-        resolver.resolve("end", payload.end.match, anchor=True), payload.end.quantifier
-    )
-    forbidden = _side_readings(resolver.resolve("forbidden", payload.forbidden, anchor=False), None)
-    return _decide(
-        (
-            _nothing_inside(_between_window(start, end), blocked)
-            for start in starts
-            for end in ends
-            for blocked in forbidden
-        ),
-        on_missing,
-    )
-
-
-def _prefix_window(anchor: _Reading) -> tuple[int, int] | None:
-    """``[0, anchor)`` as an exclusive interval, or ``None`` where nothing anchors it."""
-    return (-1, anchor[0]) if anchor else None
-
-
-def _between_window(start: _Reading, end: _Reading) -> tuple[int, int] | None:
-    """``(start, end)``, or ``None`` where the trial holds no such window.
-
-    An inverted or empty window is unmatched rather than vacuously satisfied: the
-    author's anchors did not occur in the declared order, so ``on_missing`` decides
-    and defaults to a named failure.
-    """
-    if not start or not end or start[0] >= end[0]:
-        return None
-    return (start[0], end[0])
-
-
-def _nothing_inside(window: tuple[int, int] | None, forbidden: _Reading) -> bool | None:
-    if window is None:
-        return None
-    low, high = window
-    return not any(low < position < high for position in forbidden)
-
-
-def _all_of(
-    payload: list[TraceConstraintExpr], resolver: _Resolver, on_missing: OnMissing
-) -> _Truth:
-    return _conjunction(_evaluate(expr, resolver, on_missing) for expr in payload)
-
-
-def _any_of(
-    payload: list[TraceConstraintExpr], resolver: _Resolver, on_missing: OnMissing
-) -> _Truth:
-    return _disjunction(_evaluate(expr, resolver, on_missing) for expr in payload)
-
-
-def _negate(payload: TraceConstraintExpr, resolver: _Resolver, on_missing: OnMissing) -> _Truth:
-    return _NEGATED[_evaluate(payload, resolver, on_missing)]
-
-
-def _conjunction(verdicts: Iterable[_Truth]) -> _Truth:
-    seen = set(verdicts)
-    if _Truth.FALSE in seen:
-        return _Truth.FALSE
-    if _Truth.WITHHELD in seen:
-        return _Truth.WITHHELD
-    return _Truth.UNKNOWN if _Truth.UNKNOWN in seen else _Truth.TRUE
-
-
-def _disjunction(verdicts: Iterable[_Truth]) -> _Truth:
-    seen = set(verdicts)
-    if _Truth.TRUE in seen:
-        return _Truth.TRUE
-    if _Truth.WITHHELD in seen:
-        return _Truth.WITHHELD
-    return _Truth.UNKNOWN if _Truth.UNKNOWN in seen else _Truth.FALSE
-
-
-_NEGATED: Mapping[_Truth, _Truth] = {
-    _Truth.TRUE: _Truth.FALSE,
-    _Truth.FALSE: _Truth.TRUE,
-    _Truth.UNKNOWN: _Truth.UNKNOWN,
-    _Truth.WITHHELD: _Truth.WITHHELD,
-}
-
-_HANDLERS: Mapping[TraceConstraintKind, Callable[[Any, _Resolver, OnMissing], _Truth]] = {
-    TraceConstraintKind.PRESENT: _present,
-    TraceConstraintKind.ABSENT: _absent,
-    TraceConstraintKind.COUNT: _count,
-    TraceConstraintKind.BEFORE: _before,
-    TraceConstraintKind.IMMEDIATELY_BEFORE: _immediately_before,
-    TraceConstraintKind.ABSENT_BEFORE: _absent_before,
-    TraceConstraintKind.ABSENT_BETWEEN: _absent_between,
-    TraceConstraintKind.ALL_OF: _all_of,
-    TraceConstraintKind.ANY_OF: _any_of,
-    TraceConstraintKind.NEGATE: _negate,
-}
-
-# One reading of one side: the positions it holds under one completion of the
-# undecidable evidence. Empty means the side matched nothing, which is an
-# unmatched constraint rather than a verdict.
-_Reading = tuple[int, ...]
-
-
-def _side_readings(
-    outcome: MatcherOutcome, quantifier: Quantifier | AnchorQuantifier | None
-) -> list[_Reading]:
-    """Every position set this side could hold once the missing evidence is known.
-
-    Two readings bound a side quantified ``any`` or ``all`` **once some event
-    definitely matches it**: existential quantification is monotone in the matched
-    set and universal quantification antitone, whatever relation the constraint
-    applies, so the definitely-matched set and the everything-matched set bracket
-    every completion between them.
-
-    With nothing definitely matched the bracket breaks, because its bottom is the
-    empty reading and an empty side is not a vacuous quantification: it is an
-    unmatched anchor, whose verdict ``on_missing`` supplies rather than the
-    relation. A reading of one undecidable event can therefore fall outside the two
-    ends — the singletons are where a universal reading is largest and an
-    existential one smallest — so they are enumerated alongside them.
-
-    A ``first`` / ``last`` side instead **selects** one event, and every
-    undecidable event ahead of the earliest (behind the latest) definite match is a
-    selection some completion makes. Those are enumerated rather than bracketed,
-    because two extremes bound a verdict only where the relation reading the
-    selection is monotone in it, and two of the kinds are not: an undecidable event
-    landing between two matches turns an unsatisfied ``immediately_before`` into a
-    satisfied one, and an undecidable anchor can be the one that makes an
-    ``absent_between`` window exist at all rather than widening it.
-    """
-    matched = tuple(event.position for event in outcome.matched)
-    unknown = tuple(event.position for event in outcome.undecidable)
-    if quantifier is not None and quantifier in _SELECTING_QUANTIFIERS:
-        return _selections(matched, unknown, earliest=quantifier == AnchorQuantifier.FIRST)
-    if not unknown:
-        return [matched]
-    if not matched:
-        return [(), *((position,) for position in unknown), tuple(sorted(unknown))]
-    return [matched, tuple(sorted(matched + unknown))]
-
-
-# Both quantifier domains spell these the same, and a ``str`` enum compares and
-# hashes as its value, so one set covers a ``MatcherSide`` and an ``AnchorSide``.
-_SELECTING_QUANTIFIERS = frozenset({AnchorQuantifier.FIRST, AnchorQuantifier.LAST})
-
-
-def _selections(matched: _Reading, unknown: _Reading, *, earliest: bool) -> list[_Reading]:
-    """Every single event a ``first`` / ``last`` side could end up selecting."""
-    if not matched:
-        return [(), *((position,) for position in unknown)]
-    chosen = min(matched) if earliest else max(matched)
-    ahead = [
-        position for position in unknown if (position < chosen if earliest else position > chosen)
-    ]
-    return [(position,) for position in [*ahead, chosen]]
-
-
-def _decide(values: Iterable[bool | None], on_missing: OnMissing) -> _Truth:
-    """The verdict every reachable reading agrees on, or ``UNKNOWN``.
-
-    ``None`` is an unmatched anchor — a question the trial answered by not
-    containing the anchor at all, not one the missing record left open — so
-    ``on_missing`` resolves it before the readings are compared. Under
-    ``OnMissing.WITHHOLD`` a single unmatched anchor withholds the whole reading;
-    with no unmatched anchor at all, ``WITHHOLD`` reads as ``FAIL`` — the
-    ordering or adjacency was decidable.
-    """
-    readings = list(values)
-    if on_missing is OnMissing.WITHHOLD and any(value is None for value in readings):
-        return _Truth.WITHHELD
-    unmatched_verdict = on_missing is OnMissing.PASS
-    agreed = {unmatched_verdict if value is None else value for value in readings}
-    if agreed == {True}:
-        return _Truth.TRUE
-    if agreed == {False}:
-        return _Truth.FALSE
-    return _Truth.UNKNOWN
-
-
-_VIEW_KINDS: Mapping[AdjacencyView, frozenset[TraceEventKind]] = {
-    AdjacencyView.TOOL_CALLS: frozenset({TraceEventKind.TOOL_CALL}),
-    AdjacencyView.TOOL_RESULTS: frozenset({TraceEventKind.TOOL_RESULT}),
-    AdjacencyView.MESSAGES: frozenset(
-        {TraceEventKind.ASSISTANT_MESSAGE, TraceEventKind.USER_MESSAGE}
-    ),
-    AdjacencyView.EVENTS: frozenset(TraceEventKind),
-}
-
-
-def _adjacency(timeline: TrialTimeline, among: AdjacencyView) -> Callable[[int, int], bool]:
-    """Whether one position immediately precedes another in the named view.
-
-    An event the view does not contain is adjacent to nothing: ``among:
-    tool_calls`` over a matched message names a sequence that message has no place
-    in.
-    """
-    kinds = _VIEW_KINDS[among]
-    ranks = {
-        event.position: rank
-        for rank, event in enumerate(event for event in timeline.events if event.kind in kinds)
-    }
-
-    def immediately_precedes(left: int, right: int) -> bool:
-        if left not in ranks or right not in ranks:
-            return False
-        return ranks[right] == ranks[left] + 1
-
-    return immediately_precedes
-
-
-# Why a definite failure failed, per kind — the sentence a task author reads
-# beside the constraint's id in the grade.
-_FAILURE_DETAIL: Mapping[TraceConstraintKind, str] = {
-    TraceConstraintKind.PRESENT: "no event matched",
-    TraceConstraintKind.ABSENT: "an event matched",
-    TraceConstraintKind.COUNT: "the number of matching events is outside the declared bounds",
-    TraceConstraintKind.BEFORE: (
-        "no match is ordered before the other side under the declared quantifiers"
-    ),
-    TraceConstraintKind.IMMEDIATELY_BEFORE: (
-        "no match is immediately followed by the other side in that view"
-    ),
-    TraceConstraintKind.ABSENT_BEFORE: "a forbidden event occurs before the anchor",
-    TraceConstraintKind.ABSENT_BETWEEN: "a forbidden event occurs inside the window",
-    TraceConstraintKind.ALL_OF: "a nested expression does not hold",
-    TraceConstraintKind.ANY_OF: "no nested expression holds",
-    TraceConstraintKind.NEGATE: "the negated expression holds",
-}
-
-
-def _message(
-    truth: _Truth, kind: TraceConstraintKind, resolver: _Resolver, on_missing: OnMissing
-) -> str:
-    """What the grade says about this constraint — empty when it passed."""
-    if truth is _Truth.TRUE:
-        return ""
-    if truth is _Truth.UNKNOWN:
-        return f"{kind.value} cannot be decided — " + "; ".join(resolver.undecided())
-    if truth is _Truth.WITHHELD:
-        empty = resolver.empty_sides()
-        return f"{kind.value} withheld: {' and '.join(empty)} selected no event"
-    unmatched = resolver.unmatched_anchors() if on_missing is OnMissing.FAIL else []
-    if unmatched:
-        return f"{kind.value} is unmatched: {' and '.join(unmatched)} selected no event"
-    return f"{kind.value}: {_FAILURE_DETAIL[kind]}"
